Extract duplicate `HTTPClientSpy` into the shared scope.
