+++ EssentialFeed/EssentialFeedTests/Feed API/LoadFeedFromRemoteUseCaseTests.swift
@@ -167,42 +167,6 @@ private extension LoadFeedFromRemoteUseCaseTests {
     }
     
     func failure(_ error: RemoteFeedLoader.Error) -> RemoteFeedLoader.Result {
-        return .failure(error)
-    }
-    
-    final class HTTPClientSpy: HTTPClient {
-
-        private struct Task: HTTPClientTask {
-            func cancel() {}
-        }
-
-        private var messages: [(url: URL, completion: (HTTPClient.Result) -> Void)] = []
-        
-        var requestedURLs: [URL] {
-            messages.map { $0.url }
-        }
-        
-        func get(from url: URL, completion: @escaping (HTTPClient.Result) -> Void) -> HTTPClientTask {
-            messages.append((url, completion))
-            return Task()
-        }
-        
-        func complete(with error: Error, at index: Int = 0) {
-            messages[index].completion(.failure(error))
-        }
-        
-        func complete(
-            withStatusCode code: Int,
-            data: Data,
-            at index: Int = 0
-        ) {
-            let response = HTTPURLResponse(
-                url: requestedURLs[index],
-                statusCode: code,
-                httpVersion: nil,
-                headerFields: nil
-            )!
-            messages[index].completion(.success((data, response)))
-        }
+        .failure(error)
     }
 }
